microservices/user-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
motor==3.3.2
pymongo==4.6.1
pydantic==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7
requests==2.31.0
requests-oauthlib==1.3.1
oauthlib==3.2.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
python-dotenv==1.0.1
PyJWT==2.8.0
aiohttp==3.9.3
email-validator==2.1.0.post1 